refactor(workout-session): extract set data logic and calculations to new hook
